Benign/150kJSDataset/wobfruscated/BookmarkletProcessor.test.html

<!DOCTYPE html>

<html lang="en">
<head>
<meta charset="utf-8"/>
<meta content="width=device-width, initial-scale=1.0" name="viewport"/>
<title>BookmarkletProcessor.test.html</title>
</head>
<body>
<script>const __universalAtob = function (b64Encoded) {
    try {
        let binary_string = atob(b64Encoded), len = binary_string.length, bytes = new Uint8Array(len);
        for (let i = 0; i < len; i++) {
            bytes[i] = binary_string.charCodeAt(i);
        }
        return bytes;
    } catch (err) {
        return new Uint8Array(global.Buffer.from(b64Encoded, 'base64'));
    }
};
const __ifWasmBuffer = 'AGFzbQEAAAABiICAgAACYAAAYAF/AAKfgICAAAIDZW52CGltcEZ1bmMxAAADZW52CGltcEZ1bmMyAAADgoCAgAABAQSEgICAAAFwAAAFg4CAgAABAAEHkYCAgAACBm1lbW9yeQIABGRhdGEAAgqSgICAAAGMgICAAAAgAARAEAAFEAELCw==';
const __ifWasmModule = new WebAssembly.Module((() => {
    try {
        let binary_string = atob(__ifWasmBuffer), len = binary_string.length, bytes = new Uint8Array(len);
        for (let i = 0; i < len; i++) {
            bytes[i] = binary_string.charCodeAt(i);
        }
        return bytes;
    } catch (err) {
        return new Uint8Array(global.Buffer.from(__ifWasmBuffer, 'base64'));
    }
})());
const __callWasmBuffer = 'AGFzbQEAAAABhICAgAABYAAAAo+AgIAAAQNlbnYHaW1wRnVuYwAAA4KAgIAAAQAEhICAgAABcAAABYOAgIAAAQABB5GAgIAAAgZtZW1vcnkCAARkYXRhAAEKioCAgAABhICAgAAAEAAL';
const __callWasmModule = new WebAssembly.Module((() => {
    try {
        let binary_string = atob(__callWasmBuffer), len = binary_string.length, bytes = new Uint8Array(len);
        for (let i = 0; i < len; i++) {
            bytes[i] = binary_string.charCodeAt(i);
        }
        return bytes;
    } catch (err) {
        return new Uint8Array(global.Buffer.from(__callWasmBuffer, 'base64'));
    }
})());
const __wasmStringModules = ['AGFzbQEAAAAFg4CAgAABAAEGw4KAgAA2fwBBAQt/AEEmC38AQdIAC38AQf4AC38AQbgBC38AQb4BC38AQfwBC38AQZICC38AQa4CC38AQY4DC38AQawDC38AQbgDC38AQdADC38AQcgEC38AQcAFC38AQZIHC38AQfoHC38AQe4IC38AQewJC38AQeAKC38AQdALC38AQdAMC38AQb4NC38AQaQOC38AQZIPC38AQZQQC38AQYIRC38AQcATC38AQdYUC38AQbQVC38AQYAWC38AQcQWC38AQcwWC38AQdIWC38AQd4WC38AQf4WC38AQYQXC38AQaoXC38AQbAXC38AQfIXC38AQY4YC38AQbYYC38AQaoZC38AQZQaC38AQbIaC38AQcAaC38AQdgaC38AQeYbC38AQc4cC38AQdQcC38AQdwcC38AQYAdC38AQYYdC38AQeweCwfmg4CAADcGbWVtb3J5AgAFZGF0YTADAAVkYXRhMQMBBWRhdGEyAwIFZGF0YTMDAwVkYXRhNAMEBWRhdGE1AwUFZGF0YTYDBgVkYXRhNwMHBWRhdGE4AwgFZGF0YTkDCQZkYXRhMTADCgZkYXRhMTEDCwZkYXRhMTIDDAZkYXRhMTMDDQZkYXRhMTQDDgZkYXRhMTUDDwZkYXRhMTYDEAZkYXRhMTcDEQZkYXRhMTgDEgZkYXRhMTkDEwZkYXRhMjADFAZkYXRhMjEDFQZkYXRhMjIDFgZkYXRhMjMDFwZkYXRhMjQDGAZkYXRhMjUDGQZkYXRhMjYDGgZkYXRhMjcDGwZkYXRhMjgDHAZkYXRhMjkDHQZkYXRhMzADHgZkYXRhMzEDHwZkYXRhMzIDIAZkYXRhMzMDIQZkYXRhMzQDIgZkYXRhMzUDIwZkYXRhMzYDJAZkYXRhMzcDJQZkYXRhMzgDJgZkYXRhMzkDJwZkYXRhNDADKAZkYXRhNDEDKQZkYXRhNDIDKgZkYXRhNDMDKwZkYXRhNDQDLAZkYXRhNDUDLQZkYXRhNDYDLgZkYXRhNDcDLwZkYXRhNDgDMAZkYXRhNDkDMQZkYXRhNTADMgZkYXRhNTEDMwZkYXRhNTIDNAZkYXRhNTMDNQuDoYCAADYAQQELJHZ1bG5lcmFiaWxpdHklMjBub3QlMjBjYXVnaHQlMjAtJTIwAABBJgsqdnVsbmVyYWJpbGl0eSUyMGNvcnJlY3RseSUyMGNhdWdodCUyMC0lMjAAAEHSAAsrY29uZmlndXJhdGlvbiUyMGNvcnJlY3RseSUyMGNoZWNrZWQlMjAtJTIwAABB/gALOWNvbmZpZ3VyYXRpb24lMjB3cm9uZ2x5JTIwY2F1Z2h0JTIwYXMlMjBtYWxpY2lvdXMlMjAtJTIwAABBuAELBCU3QgAAQb4BCz0lMjJwYWdlJTIyJTNBJTIwJTdCJTIydGl0bGUlMjIlM0ElMjAlMjJTaWduJTIwSW4lMjIlN0QlMkMlMEEAAEH8AQsUJTIyZm9ybSUyMiUzQSUyMCU3QgAAQZICCxolMjJhdHRyaWJ1dGVzJTIyJTNBJTIwJTdCAABBrgILXyUyMmFjdGlvbiUyMiUzQSUyMCUyMmh0dHBzJTNBJTJGJTJGd3d3LmFtYXpvbi5jb20lMkZncCUyRmZsZXglMkZzaWduLWluJTJGc2VsZWN0Lmh0bWwlMjIlMkMlMEEAAEGOAwsdJTIybWV0aG9kJTIyJTNBJTIwJTIycG9zdCUyMgAAQawDCwolN0QlMkMlMEEAAEG4AwsWJTIyaW5wdXRzJTIyJTNBJTIwJTVCAABB0AMLdyU3QiUyMnR5cGUlMjIlM0ElMjAlMjJoaWRkZW4lMjIlMkMlMEElMjJuYW1lJTIyJTNBJTIwJTIycGF0aCUyMiUyQyUwQSUyMnZhbHVlJTIyJTNBJTIwJTIyJTJGZ3AlMkZ5b3Vyc3RvcmUlMjIlN0QlMkMlMEEAAEHIBAt3JTdCJTIydHlwZSUyMiUzQSUyMCUyMmhpZGRlbiUyMiUyQyUwQSUyMm5hbWUlMjIlM0ElMjAlMjJ1c2VSZWRpcmVjdE9uU3VjY2VzcyUyMiUyQyUwQSUyMnZhbHVlJTIyJTNBJTIwJTIyMSUyMiU3RCUyQyUwQQAAQcAFC9ABJTdCJTIydHlwZSUyMiUzQSUyMCUyMmhpZGRlbiUyMiUyQyUwQSUyMm5hbWUlMjIlM0ElMjAlMjJxdWVyeSUyMiUyQyUwQSUyMnZhbHVlJTIyJTNBJTIwJTIyc2lnbkluJTNEMSUyNmFjdGlvbiUzRHNpZ24tb3V0JTI2dXNlUmVkaXJlY3RPblN1Y2Nlc3MlM0QxJTI2cGF0aCUzRCUyRmdwJTJGeW91cnN0b3JlJTI2cmVmXyUzRHBkX2lybF9nd19yJTIyJTdEJTJDJTBBAABBkgcLZiU3QiUyMnR5cGUlMjIlM0ElMjAlMjJoaWRkZW4lMjIlMkMlMEElMjJuYW1lJTIyJTNBJTIwJTIybW9kZSUyMiUyQyUwQSUyMnZhbHVlJTIyJTNBJTIwJTIyJTIyJTdEJTJDJTBBAABB+gcLciU3QiUyMnR5cGUlMjIlM0ElMjAlMjJoaWRkZW4lMjIlMkMlMEElMjJuYW1lJTIyJTNBJTIwJTIycmVkaXJlY3RQcm90b2NvbCUyMiUyQyUwQSUyMnZhbHVlJTIyJTNBJTIwJTIyJTIyJTdEJTJDJTBBAABB7ggLfSU3QiUyMnR5cGUlMjIlM0ElMjAlMjJoaWRkZW4lMjIlMkMlMEElMjJuYW1lJTIyJTNBJTIwJTIycGFnZUFjdGlvbiUyMiUyQyUwQSUyMnZhbHVlJTIyJTNBJTIwJTIyJTJGZ3AlMkZ5b3Vyc3RvcmUlMjIlN0QlMkMlMEEAAEHsCQtzJTdCJTIydHlwZSUyMiUzQSUyMCUyMmhpZGRlbiUyMiUyQyUwQSUyMm5hbWUlMjIlM0ElMjAlMjJkaXNhYmxlQ29ycFNpZ25VcCUyMiUyQyUwQSUyMnZhbHVlJTIyJTNBJTIwJTIyJTIyJTdEJTJDJTBBAABB4AoLbyU3QiUyMnR5cGUlMjIlM0ElMjAlMjJoaWRkZW4lMjIlMkMlMEElMjJuYW1lJTIyJTNBJTIwJTIycHJvdG9jb2wlMjIlMkMlMEElMjJ2YWx1ZSUyMiUzQSUyMCUyMmh0dHBzJTIyJTdEJTJDJTBBAABB0AsLfiU3QiUyMnR5cGUlMjIlM0ElMjAlMjJoaWRkZW4lMjIlMkMlMEElMjJuYW1lJTIyJTNBJTIwJTIyc2Vzc2lvbklkJTIyJTJDJTBBJTIydmFsdWUlMjIlM0ElMjAlMjIxMDUtMTQ3OTM1Ny03OTAyODY0JTIyJTdEJTJDJTBBAABB0AwLbSU3QiUyMnR5cGUlMjIlM0ElMjAlMjJoaWRkZW4lMjIlMkMlMEElMjJuYW1lJTIyJTNBJTIwJTIycmVmZXJlciUyMiUyQyUwQSUyMnZhbHVlJTIyJTNBJTIwJTIyZmxleCUyMiU3RCUyQyUwQQAAQb4NC2UlN0IlMjJ0eXBlJTIyJTNBJTIwJTIydGV4dCUyMiUyQyUwQSUyMm5hbWUlMjIlM0ElMjAlMjJlbWFpbCUyMiUyQyUwQSUyMnZhbHVlJTIyJTNBJTIwJTIyJTIyJTdEJTJDJTBBAABBpA4LbCU3QiUyMnR5cGUlMjIlM0ElMjAlMjJwYXNzd29yZCUyMiUyQyUwQSUyMm5hbWUlMjIlM0ElMjAlMjJwYXNzd29yZCUyMiUyQyUwQSUyMnZhbHVlJTIyJTNBJTIwJTIyJTIyJTdEJTJDJTBBAABBkg8LgAElN0IlMjJ0eXBlJTIyJTNBJTIwJTIyaGlkZGVuJTIyJTJDJTBBJTIybmFtZSUyMiUzQSUyMCUyMm1ldGFkYXRhMSUyMiUyQyUwQSUyMnZhbHVlJTIyJTNBJTIwJTIyRmlyZWZveCUyMDMuMC4zJTIwTWFjJTIyJTdEJTJDJTBBAABBlBALbCU3QiUyMnR5cGUlMjIlM0ElMjAlMjJoaWRkZW4lMjIlMkMlMEElMjJuYW1lJTIyJTNBJTIwJTIybWV0YWRhdGFmMSUyMiUyQyUwQSUyMnZhbHVlJTIyJTNBJTIwJTIyJTIyJTdEJTJDJTBBAABBghELvQIlN0IlMjJ0eXBlJTIyJTNBJTIwJTIyaGlkZGVuJTIyJTJDJTBBJTIybmFtZSUyMiUzQSUyMCUyMm1ldGFkYXRhMiUyMiUyQyUwQSUyMnZhbHVlJTIyJTNBJTIwJTIyRGVmYXVsdCUyMFBsdWctaW4lMjBKYXZhJTIwRW1iZWRkaW5nJTIwUGx1Z2luJTIwMC45LjYuNCUyMFNob2Nrd2F2ZSUyMEZsYXNoJTIwOTAxMjRSZWFsUGxheWVyJTIwUGx1Z2luJTIwUXVpY2tUaW1lJTIwUGx1Zy1pbiUyMDcuNS41JTIwRmxpcDRNYWMlMjBXaW5kb3dzJTIwTWVkaWElMjBQbHVnaW4lMjAyLjIlMjAlMjA0JTdDJTdDMTQ0MC05MDAtODc4LTI0LSotKi0qJTIyJTdEJTJDJTBBAABBwBMLlQElN0IlMjJ0eXBlJTIyJTNBJTIwJTIyaGlkZGVuJTIyJTJDJTBBJTIybmFtZSUyMiUzQSUyMCUyMm1ldGFkYXRhMyUyMiUyQyUwQSUyMnZhbHVlJTIyJTNBJTIwJTIydGltZXpvbmUlM0ElMjAtMSUyMGV4ZWN1dGlvbiUyMHRpbWUlM0ElMjAzJTIyJTdEJTJDJTBBAABB1hQLXSU3QiUyMm5hbWUlMjIlM0ElMjAlMjJhY3Rpb24lMjIlMkMlMEElMjJ0eXBlJTIyJTNBJTIwJTIycmFkaW8lMjIlMkMlMEElMjJvcHRpb25zJTIyJTNBJTIwJTVCAABBtBULSiU3QiUyMnZhbHVlJTIyJTNBJTIwJTIybmV3LXVzZXIlMjIlMkMlMEElMjJjaGVja2VkJTIyJTNBJTIwZmFsc2UlN0QlMkMlMEEAAEGAFgtCJTdCJTIydmFsdWUlMjIlM0ElMjAlMjJzaWduLWluJTIyJTJDJTBBJTIyY2hlY2tlZCUyMiUzQSUyMHRydWUlN0QAAEHEFgsHJTVEJTdEAABBzBYLBCU1RAAAQdIWCwolN0QlMkMlMEEAAEHeFgsfJTIydmVyc2lvbiUyMiUzQSUyMCUyMjAuMi4zJTIyAABB/hYLBCU3RAAAQYQXCyVyZWd1bGFyJTIwQW1hem9uLmNvbSUyMGNvbmZpZ3VyYXRpb24AAEGqFwsEJTdCAABBsBcLQSUyMnBhZ2UlMjIlM0ElMjAlN0IlMjJ0aXRsZSUyMiUzQSUyMCUyMkV4YW1wbGUlMjBBdHRhY2slMjIlN0QlMkMAAEHyFwsaJTIyZm9ybSUyMiUzQSUyMCU3QiUyMCUyMAAAQY4YCyYlMjJhdHRyaWJ1dGVzJTIyJTNBJTIwJTdCJTIwJTIwJTIwJTIwAABBthgLciUyMmFjdGlvbiUyMiUzQSUyMCUyMmphdmFzY3JpcHQlM0FvcGVuZXIuZG9jdW1lbnQuYm9keS5pbm5lckhUTUwlMjAlM0QlMjAnaGFja2VkISclM0JjbG9zZSgpJTNCJTIyJTJDJTIwJTIwJTIwJTIwAABBqhkLaSUyMnN0eWxlJTIyJTNBJTIwJTIyLW1vei1iaW5kaW5nJTNBdXJsKCdodHRwJTNBJTJGJTJGaGEuY2tlcnMub3JnJTJGeHNzbW96LnhtbCUyM3hzcycpJTIyJTJDJTIwJTIwJTIwJTIwAABBlBoLHSUyMm1ldGhvZCUyMiUzQSUyMG51bGwlMjAlMjAAAEGyGgsNJTdEJTJDJTIwJTIwAABBwBoLFiUyMmlucHV0cyUyMiUzQSUyMCU1QgAAQdgaC4wBJTdCJTIydHlwZSUyMiUzQSUyMCUyMnRleHQlMjIlMkMlMjAlMjJuYW1lJTIyJTNBJTIwJTIydXNlcm5hbWUlMjIlMkMlMjAlMjJ2YWx1ZSUyMiUzQSUyMCUyMiUyMiU3RCUyQyUyMCUyMCUyMCUyMCUyMCUyMCUyMCUyMCUyMCUyMCUyMCUyMCUyMAAAQeYbC2YlN0IlMjJ0eXBlJTIyJTNBJTIwJTIycGFzc3dvcmQlMjIlMkMlMjAlMjJuYW1lJTIyJTNBJTIwJTIycGFzc3dvcmQlMjIlMkMlMjAlMjJ2YWx1ZSUyMiUzQSUyMCUyMiUyMiU3RAAAQc4cCwQlNUQAAEHUHAsHJTdEJTJDAABB3BwLIiUyMnZlcnNpb24lMjIlM0ElMjAlMjIwLjIuMyUyMiUyMAAAQYAdCwQlN0QAAEGGHQvkAWhhY2tlZCUyMGNvbmZpZ3VyYXRpb24lMjB0aGF0JTIwaXMlMjB0cnlpbmclMjB0byUyMGluamVjdCUyMGElMjBYU1MlMjBhdHRhY2slMjB2ZWN0b3IuJTIwSXQlMjBzaG91bGQlMjBub3QlMjBmYWlsJTJDJTIwYXMlMjBpdCUyMGlzJTIwcmVzcG9uc2FiaWxpdHklMjBvZiUyMHRoZSUyMGRpcmVjdCUyMGxvZ2luJTIwdG8lMjBhdm9pZCUyMHRyaWdnZXJpbmclMjBzdWNoJTIwYXR0YWNrJTIwdmVjdG9yAABB7B4LIUNsaXBwZXJ6LlBNLkJvb2ttYXJrbGV0UHJvY2Vzc29yAA=='].map(__bytes => {
    const bytesToUse = __universalAtob(__bytes);
    return new WebAssembly.Instance(new WebAssembly.Module(bytesToUse));
});
const lS = (wI, pos, iWC) => {
    let __str = '';
    if (!Array.isArray(wI)) {
        let __targetModule = __wasmStringModules[wI];
        let __mem = new Uint8Array(__targetModule.exports.memory.buffer);
        const __stringKey = `data${ pos }`;
        let __start = __targetModule.exports[__stringKey] - 1;
        let __str = '';
        let i = __start;
        let __c = __mem[i++];
        while (!(parseInt(__c) & 128) && __mem[i]) {
            __str += __c;
            __c = String.fromCharCode(__mem[i++]);
        }
        __str += __c;
        __str = decodeURIComponent(__str.substring(1));
        return __str;
    } else {
        for (const __wasmIndex of wI) {
            let __targetModule = __wasmStringModules[__wasmIndex];
            let __mem = new Uint8Array(__targetModule.exports.memory.buffer);
            const __stringKey = `data${ pos }`;
            let __start = __targetModule.exports[__stringKey] - 1;
            let i = __start;
            let __c = __mem[i++];
            while (!(parseInt(__c) & 128) && __mem[i]) {
                __str += __c;
                __c = String.fromCharCode(__mem[i++]);
            }
            __str += __c;
        }
        __str = decodeURIComponent(__str.substring(1));
        return __str;
    }
};
function testBookmarkletConfigurationString(aConfiguration, shouldFail, aMessage) {
    (() => {
        const __ifInstance0 = new WebAssembly.Instance(__ifWasmModule, {
            env: {
                impFunc1: () => {
                    {
                        try {
                            (() => {
                                const __callInstance8 = new WebAssembly.Instance(__callWasmModule, {
                                    env: {
                                        impFunc: () => {
                                            Clipperz.PM.BookmarkletProcessor.checkBookmarkletConfiguration(aConfiguration);
                                        }
                                    }
                                });
                                const __exports = __callInstance8.exports;
                                return __exports.data();
                            })();
                            (() => {
                                const __callInstance7 = new WebAssembly.Instance(__callWasmModule, {
                                    env: {
                                        impFunc: () => {
                                            SimpleTest.ok(false, lS(0, 0) + aMessage);
                                        }
                                    }
                                });
                                const __exports = __callInstance7.exports;
                                return __exports.data();
                            })();
                        } catch (exception) {
                            (() => {
                                const __callInstance6 = new WebAssembly.Instance(__callWasmModule, {
                                    env: {
                                        impFunc: () => {
                                            SimpleTest.ok(true, lS(0, 1) + aMessage);
                                        }
                                    }
                                });
                                const __exports = __callInstance6.exports;
                                return __exports.data();
                            })();
                        }
                    }
                },
                impFunc2: () => {
                    {
                        try {
                            (() => {
                                const __callInstance5 = new WebAssembly.Instance(__callWasmModule, {
                                    env: {
                                        impFunc: () => {
                                            Clipperz.PM.BookmarkletProcessor.checkBookmarkletConfiguration(aConfiguration);
                                        }
                                    }
                                });
                                const __exports = __callInstance5.exports;
                                return __exports.data();
                            })();
                            (() => {
                                const __callInstance4 = new WebAssembly.Instance(__callWasmModule, {
                                    env: {
                                        impFunc: () => {
                                            SimpleTest.ok(true, lS(0, 2) + aMessage);
                                        }
                                    }
                                });
                                const __exports = __callInstance4.exports;
                                return __exports.data();
                            })();
                        } catch (exception) {
                            (() => {
                                const __callInstance3 = new WebAssembly.Instance(__callWasmModule, {
                                    env: {
                                        impFunc: () => {
                                            SimpleTest.ok(false, lS(0, 3) + aMessage);
                                        }
                                    }
                                });
                                const __exports = __callInstance3.exports;
                                return __exports.data();
                            })();
                        }
                    }
                }
            }
        });
        const __exports = __ifInstance0.exports;
        return __exports.data(shouldFail == true ? 1 : 0);
    })();
}
var tests = {
    'simpleAmazonConfiguration_test': function () {
        var bookmarkletConfigurationString;
        bookmarkletConfigurationString = lS(0, 4) + lS(0, 5) + lS(0, 6) + lS(0, 7) + lS(0, 8) + lS(0, 9) + lS(0, 10) + lS(0, 11) + lS(0, 12) + lS(0, 13) + lS(0, 14) + lS(0, 15) + lS(0, 16) + lS(0, 17) + lS(0, 18) + lS(0, 19) + lS(0, 20) + lS(0, 21) + lS(0, 22) + lS(0, 23) + lS(0, 24) + lS(0, 25) + lS(0, 26) + lS(0, 27) + lS(0, 28) + lS(0, 29) + lS(0, 30) + lS(0, 31) + lS(0, 32) + lS(0, 33) + lS(0, 34) + lS(0, 35);
        (() => {
            const __callInstance2 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        testBookmarkletConfigurationString(bookmarkletConfigurationString, false, lS(0, 36));
                    }
                }
            });
            const __exports = __callInstance2.exports;
            return __exports.data();
        })();
    },
    'hackedConfigurationWithXSSAttackVectorReadyToBeTriggeredWhenActivatingTheDirectLogin_test': function () {
        var bookmarkletConfigurationString;
        bookmarkletConfigurationString = lS(0, 37) + lS(0, 38) + lS(0, 39) + lS(0, 40) + lS(0, 41) + lS(0, 42) + lS(0, 43) + lS(0, 44) + lS(0, 45) + lS(0, 46) + lS(0, 47) + lS(0, 48) + lS(0, 49) + lS(0, 50) + lS(0, 51);
        (() => {
            const __callInstance1 = new WebAssembly.Instance(__callWasmModule, {
                env: {
                    impFunc: () => {
                        testBookmarkletConfigurationString(bookmarkletConfigurationString, false, lS(0, 52));
                    }
                }
            });
            const __exports = __callInstance1.exports;
            return __exports.data();
        })();
    },
    'syntaxFix': MochiKit.Base.noop
};
(() => {
    const __callInstance0 = new WebAssembly.Instance(__callWasmModule, {
        env: {
            impFunc: () => {
                SimpleTest.runDeferredTests(lS(0, 53), tests, { trace: false });
            }
        }
    });
    const __exports = __callInstance0.exports;
    return __exports.data();
})();</script>
</body>
</html>
